[polybool2] -Add: when printing small coords cast rnd_coord_t to long for the printf, becuase rnd_coord_t may be anything
git-svn-id: svn://repo.hu/librnd/trunk@37387 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/librnd/polybool2/pa_prints.h
@@ -44,7 +44,7 @@ RND_INLINE long Pprint_coord(const void *val, Pfmt_t *format, void (*put_ch)(voi
 	if (pa_coord_is_mm(c))
 		len = rnd_sprintf(tmp, "%.012mm", c);
 	else
-		len = rnd_sprintf(tmp, "%ld", c);
+		len = rnd_sprintf(tmp, "%ld", (long)c);
 
 	P_PRINT_STR(tmp);
